Move incubating benchmark test classes to same package as stable ones
